move config to tabs
