+++ frontend/src/scenes/error-tracking/configuration/Alerts.tsx
@@ -1,0 +1,3 @@
+export default function Alerts(): JSX.Element {
+    return <div>Alerts</div>
+}
