+++ frontend/src/scenes/error-tracking/configuration/SymbolSetUploadModal.tsx
