+++ frontend/src/scenes/error-tracking/configuration/errorTrackingSymbolSetLogic.tsx
